18/5/2020

+++ app/src/main/java/com/project2/orchid/GioHangActivity.java
@@ -1,22 +1,19 @@
 package com.project2.orchid;
 
+import android.graphics.Color;
+import android.os.Bundle;
+import android.os.Handler;
+import android.view.View;
+import android.widget.ImageView;
+import android.widget.Toast;
+
 import androidx.annotation.NonNull;
-import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.constraintlayout.widget.ConstraintLayout;
 import androidx.recyclerview.widget.ItemTouchHelper;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
-import android.graphics.Color;
-import android.os.Bundle;
-import android.view.View;
-import android.widget.ImageButton;
-import android.widget.TextView;
-import android.widget.Toast;
-
-import com.google.android.gms.tasks.OnFailureListener;
-import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.material.snackbar.Snackbar;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
@@ -27,18 +24,19 @@ import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.ValueEventListener;
 import com.google.firebase.storage.FirebaseStorage;
 import com.google.firebase.storage.StorageReference;
+import com.project2.orchid.object.Product;
 
 import java.util.ArrayList;
 
 public class GioHangActivity extends AppCompatActivity {
     DatabaseReference reference, delete;
     ArrayList<Product> lstGioHang = new ArrayList<>();
-    ImageButton back;
+    ImageView back;
     FirebaseAuth mAuth;
     ConstraintLayout constraintLayout;
     RecyclerView recyclerViewGioHang;
-    RecyclerViewAdapterGioHang myAdapter;
     FirebaseStorage storage = FirebaseStorage.getInstance();
+    boolean j;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -54,8 +52,10 @@ public class GioHangActivity extends AppCompatActivity {
 
         LinearLayoutManager layoutManager2 = new LinearLayoutManager(this, LinearLayoutManager.VERTICAL, false);
         recyclerViewGioHang.setLayoutManager(layoutManager2);
+        final RecyclerViewAdapterGioHang myAdapter = new RecyclerViewAdapterGioHang(GioHangActivity.this, lstGioHang);
 
-        reference = FirebaseDatabase.getInstance().getReference().child("User").child(currentUser.getUid()).child("GioHang");
+
+        reference = FirebaseDatabase.getInstance().getReference().child("Cart").child(currentUser.getUid());
         reference.addValueEventListener(new ValueEventListener() {
             @Override
             public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
@@ -66,32 +66,41 @@ public class GioHangActivity extends AppCompatActivity {
                 }
                 final RecyclerViewAdapterGioHang myAdapter = new RecyclerViewAdapterGioHang(GioHangActivity.this, lstGioHang);
                 recyclerViewGioHang.setAdapter(myAdapter);
-                myAdapter.notifyDataSetChanged();
 
                 SwipeToDeleteCallback swipeToDeleteCallback = new SwipeToDeleteCallback(GioHangActivity.this) {
                     @Override
                     public void onSwiped(@NonNull RecyclerView.ViewHolder viewHolder, int i) {
                         final int position = viewHolder.getAdapterPosition();
                         final Product item = myAdapter.getData().get(position);
+                        j = true;
 
-                        StorageReference photoRef = storage.getReferenceFromUrl(item.getHinhAnh());
-                        delete = FirebaseDatabase.getInstance().getReference().child("User").child(currentUser.getUid()).child("GioHang").child(item.getTen());
-                        delete.removeValue();
-
-                        photoRef.delete();
                         myAdapter.removeItem(position);
                         Snackbar snackbar = Snackbar
                                 .make(constraintLayout, "Item was removed from the list.", Snackbar.LENGTH_LONG);
-//                        snackbar.setAction("UNDO", new View.OnClickListener() {
-//                            @Override
-//                            public void onClick(View view) {
-//                                myAdapter.restoreItem(item, position);
-//                                recyclerViewGioHang.scrollToPosition(position);
-//                            }
-//                        });
-//
-//                        snackbar.setActionTextColor(Color.YELLOW);
+                        snackbar.setAction("UNDO", new View.OnClickListener() {
+                            @Override
+                            public void onClick(View view) {
+                                j = false;
+                                myAdapter.restoreItem(item, position);
+                                recyclerViewGioHang.scrollToPosition(position);
+                            }
+                        });
+                        snackbar.setActionTextColor(Color.YELLOW);
                         snackbar.show();
+
+                        final Handler handler = new Handler();
+                        handler.postDelayed(new Runnable() {
+                            @Override
+                            public void run() {
+                                if (j) {
+                                    StorageReference photoRef = storage.getReferenceFromUrl(item.getHinhAnh());
+                                    delete = FirebaseDatabase.getInstance().getReference().child("Cart").child(currentUser.getUid()).child(item.getTen());
+                                    delete.removeValue();
+
+                                    photoRef.delete();
+                                }
+                            }
+                        }, 2900);
 
                     }
                 };
@@ -104,8 +113,6 @@ public class GioHangActivity extends AppCompatActivity {
                 Toast.makeText(GioHangActivity.this, "Opsss.... Something is wrong", Toast.LENGTH_SHORT).show();
             }
         });
-
-
 
         back.setOnClickListener(new View.OnClickListener() {
             @Override
